Removed old FlowDistributor
