+++ opendc-compute/opendc-compute-simulator/src/main/kotlin/org/opendc/compute/simulator/provisioner/HostsProvisioningStep.kt
@@ -29,7 +29,7 @@ import org.opendc.compute.topology.specs.ClusterSpec
 import org.opendc.compute.topology.specs.HostSpec
 import org.opendc.simulator.compute.power.SimPowerSource
 import org.opendc.simulator.engine.engine.FlowEngine
-import org.opendc.simulator.engine.graph.FlowDistributorNew
+import org.opendc.simulator.engine.graph.FlowDistributor
 
 /**
  * A [ProvisioningStep] that provisions a list of hosts for a [ComputeService].
@@ -64,7 +64,7 @@ public class HostsProvisioningStep internal constructor(
             service.addPowerSource(simPowerSource)
             simPowerSources.add(simPowerSource)
 
-            val powerDistributor = FlowDistributorNew(graph)
+            val powerDistributor = FlowDistributor(graph)
             graph.addEdge(powerDistributor, simPowerSource)
 
             // Create hosts, they are connected to the powerMux when SimMachine is created
